==> Notebooks/2_population_density.ipynb <==
import pandas as pd
import numpy as np
import re
import seaborn as sns
import matplotlib.pyplot as plt
from bs4 import BeautifulSoup as bs
import requests

import warnings
# Establece el filtro para ignorar todos los warnings
warnings.simplefilter("ignore")

#datos del censo para la densidad de población:

poblacion = pd.read_excel('../Data/Raw/poblacion_2023.xlsx')

poblacion.head(1)

codigo = []
municipio = []
for x in poblacion['Unidad Poblacional']:
    codigo.append(x.split(' ',1)[0])
    municipio.append(x.split(' ',1)[1])
    
poblacion['código_unidad'] = codigo
poblacion['mun_nom'] = municipio
poblacion = poblacion[poblacion['código_unidad'] == '000000']
poblacion.drop(columns=['Unidad Poblacional'], inplace = True)

poblacion.reset_index(inplace = True, drop = True)

poblacion['Municipio']=poblacion['Municipio'].apply(lambda x: str(x).zfill(3))

poblacion['Provincia']=poblacion['Provincia'].apply(lambda x: str(x))

poblacion.head(1)

poblacion['Cod. INE']=poblacion['Provincia']+poblacion['Municipio']+poblacion['código_unidad']
poblacion.columns = poblacion.columns.str.replace(' ','_').str.lower().str.replace('.','')
poblacion.drop(columns=['código_unidad'], inplace = True)
poblacion.columns

test = poblacion.copy()

replace_dict = {r'(\b\w*) \(LA\)':'LA ',
                r'(\b\w*) \(EL\)':'EL ',
               r'(\b\w*) \(LAS\)':'LAS ',
               r'(\b\w*) \(LOS\)':'LOS ',
               r"(\b\w*) \(L'\)":"L'",
               r'(\b\w*) \(A\)':'A ',
               r'(\b\w*) \(O\)':'O ',
               r'(\b\w*) \(OS\)':'OS ',
                r'(\b\w*) \(AS\)':'AS ',
                'r(\b\w*) \(ELS\)': 'ELS ',
                'r(\b\w*) \(ES\)':'ES ', 
                r'(\b\w*) \(LES\)':'LES ', 
                r'(\b\w*) \(SA\)':'SA ',
                r'(\b\w*) \(SES\)':'SES '
               }

muns=[]
for x in poblacion['mun_nom']:
    if re.search(r'\b\w* \(LA\)', x, flags=re.IGNORECASE):
            mun = 'LA ' + x
            muns.append(mun)
    elif re.search(r'\b\w* \(LAS\)', x, flags=re.IGNORECASE):
            mun = 'LAS ' + x
            muns.append(mun)
    elif re.search(r'\b\w* \(LOS\)', x, flags=re.IGNORECASE):
            mun = 'LOS ' + x
            muns.append(mun)
    elif re.search(r'\b\w* \(EL\)', x, flags=re.IGNORECASE):
            mun = 'EL ' + x
            muns.append(mun)
    elif re.search(r"\b\w* \(L'\)", x, flags=re.IGNORECASE):
            mun = "L'" + x
            muns.append(mun)
    elif re.search(r'\b\w* \(A\)', x, flags=re.IGNORECASE):
            mun = 'A ' + x
            muns.append(mun)
    elif re.search(r'\b\w* \(O\)', x, flags=re.IGNORECASE):
            mun = 'O ' + x
            muns.append(mun)
    elif re.search(r'\b\w* \(AS\)', x, flags=re.IGNORECASE):
            mun = 'AS ' + x
            muns.append(mun)
    elif re.search(r'\b\w* \(OS\)', x, flags=re.IGNORECASE):
            mun = 'OS ' + x
            muns.append(mun)
    elif re.search(r'\b\w* \(ELS\)', x, flags=re.IGNORECASE):
            mun = 'ELS ' + x
            muns.append(mun)
    elif re.search(r'\b\w* \(ES\)', x, flags=re.IGNORECASE):
            mun = 'ES ' + x
            muns.append(mun)
    elif re.search(r'\b\w* \(SA\)', x, flags=re.IGNORECASE):
            mun = 'SA ' + x
            muns.append(mun)
    elif re.search(r'\b\w* \(LES\)', x, flags=re.IGNORECASE):
            mun = 'LES ' + x
            muns.append(mun)
    elif re.search(r'\b\w* \(SES\)', x, flags=re.IGNORECASE):
            mun = 'SES ' + x
            muns.append(mun)
    else:
        muns.append(x)

poblacion['municipios']=muns

poblacion.drop(columns=['mun_nom'], inplace = True)

poblacion['municipios'] = poblacion['municipios'].str.capitalize()
replace = {'á':'a', 'à':'a',
           'é':'e', 'è':'e',
           'í':'i', 'ì':'i', 
           'ó':'o', 'ò':'o', 'ô':'o', 
           'ú':'u', 
           'Á':'A', 'À':'A',
           'Ó':'O',
           'Ò':'O',
           'Ú':'U', 
           'É':'E', 
           'Ì':'I',
           'Í':'I',
           r'\([^\)]*\)':''
          }
poblacion.replace(replace, regex=True, inplace = True)
poblacion['municipios'] = poblacion['municipios'].str.strip()
poblacion['municipios'] = poblacion['municipios'].str.capitalize()
poblacion.head(3)

table_2 = pd.read_csv('../Data/Raw/superficies.csv')

sup_2 = pd.DataFrame(table_2)

sup_2.shape

sup_2.drop(columns=['Municipio#', 'Comarca', 'Altitud<br/>(m.s.n.m.)', 'PoblaciÃ³n<br/>(2019)', 'Densidad<br/>(hab./kmÂ²)'], inplace = True)

sup_2.head()

sup_2.rename(columns={'Municipio':'municipios','Superficie<br/>(kmÂ²)':'km2', 'CC.AA.':'comunidad'}, inplace = True)
sup_2.columns = sup_2.columns.str.lower().str.strip()
sup_2.columns

replacement = {'Ã¡':'a','Ã©': 'e','Ã³':'o','Ã²':'o', 'Ã¨':'e','Ã':'e', 'Ã±':'ñ','Ãº':'u','Ã':'i'}
sup_2.replace(replacement, inplace = True, regex = True)

sup_2.municipios = sup_2.municipios.str.strip()
sup_2.municipios = sup_2.municipios.str.capitalize()
sup_2.tail(2)

sup_2['km2'] = sup_2['km2'].str.replace('.', '')
sup_2['km2'] = sup_2['km2'].str.replace(',', '.')

sup_2['km2'] = sup_2['km2'].astype(float)

type(sup_2['km2'][8130])

#eliminamos más caracteres regex que luego nos dan duplicados al hacer merge (reducimos de casi mil a menos de 450)
more_replacements = {r'i\xad':'i',
                    r'I\x89':'E',
                    r'I\x93':'O',
                    r'i\x81':'a',
                    r'I\x81':'A',
                    r'I\x80':'A',
                    r'i\x80':'a',
                    r'i½':'y',
                    r'i\xa0':'a',
                    r'\([^\)]*\)':'',
                    r'i¼':'ü',
                    r'i§':'s',
                    r'i\x93':'o',
                    r'i\x9a':'u',
                    r'i\x8d':'i',
                     r'I\x8d':'I',
                    r'I\x92':'O',
                    r'I\x9a':'U',
                    'ii':'ia',
                    r'i¯':'ï',
                     'Ò':'O'
                    }

sup_2.replace(more_replacements, regex = True, inplace = True)
sup_2.municipios = sup_2.municipios.str.strip()

pop_den = pd.merge(poblacion,sup_2, how='outer', on=['municipios'])
pop_den

pop_den.isnull().sum()

pop_den.shape

#sigue habiendo duplicados, así que vamos sacando listas de reemplazo para modificar los dataframes antes de hacer merge

no_match = pop_den[pop_den['cod_ine'].isna()]
no_match.municipios.unique()

no_match_2 = pop_den[pop_den['km2'].isna()]
no_match_2.municipios.unique()

cambios = {}
cambios = pd.DataFrame(cambios)
cambios['original'] = no_match_2.municipios.unique()
cambios[['c1','c2']] = cambios['original'].str.split('/', expand = True)
cambios.c2 = cambios.c2.str.strip().str.capitalize()
cambios.head(2)

cambios_b = {}
cambios_b = pd.DataFrame(cambios_b)
cambios_b['c1'] = pd.DataFrame(no_match.municipios.unique().tolist())
cambios_b['c2'] = pd.DataFrame(no_match.municipios.unique().tolist())
cambios_b['c3'] = pd.DataFrame(no_match.municipios.unique().tolist())
cambios_b.head(2)


#merge primero con columna c2:
cambios_poblacion_1 = pd.merge(cambios, cambios_b, on='c2')
cambios_poblacion_1.head(5)

cambios_poblacion_1.drop(columns=['c1_x','c2', 'c1_y'], inplace = True)

original = cambios_poblacion_1['original'].tolist()
c3 = cambios_poblacion_1['c3'].tolist()

#merge ahora con columna c1:
cambios_poblacion_2 = pd.merge(cambios, cambios_b, on='c1')
cambios_poblacion_2.drop(columns=['c2_x','c1', 'c2_y'], inplace = True)
cambios_poblacion_2.head(2)

original.extend(cambios_poblacion_2['original'].tolist())
c3.extend(cambios_poblacion_2['c3'].tolist())

cambios_dict = dict(zip(original,c3))

poblacion.replace(cambios_dict, inplace = True)
poblacion['municipios'] = poblacion['municipios'].str.strip()
poblacion['municipios'] = poblacion['municipios'].str.capitalize()

#repetimos merge para ver con cuántos valores únicos nos quedamos ahora:
pop_den = pd.merge(poblacion,sup_2, how='outer', on=['municipios'])
pop_den.tail(5)

no_match_2 = pop_den[pop_den['km2'].isna()]
no_match_2 = no_match_2.drop_duplicates(subset=['municipios'], keep=False)
cities1 = no_match_2.municipios.unique()
no_match_2.municipios.unique()

no_match = pop_den[pop_den['cod_ine'].isna()]
no_match = no_match.drop_duplicates(subset=['municipios'], keep=False)
cities2 = no_match.municipios.unique()
no_match.municipios.unique()

#usamos la librería fuzzywuzzy para encontrar los equivalentes más cercanos a cada municipio en ambas listas para unificar nomenclaturas:


from fuzzywuzzy import process

# Listas de nombres de ciudades
cities_list1 = cities2  # Lista que queremos modificar
cities_list2 = cities1  # Lista del INE

# Comparar cada ciudad de cities_list1 con todas las ciudades de cities_list2 y reemplazar si la similitud es alta
threshold = 60  
 

replacements = {}
for city1 in cities_list1:
    match = process.extractOne(city1, cities_list2)
    if match[1] >= threshold:
        if city1 not in replacements:
            replacements[city1] = match[0]

#Revisamos y quitamos las correlaciones que no sean correctas para crear el diccionario de cambios:

replacements={
#primero los cambios sacados de fuzzywuzzy revisados:
'Samianigo': 'Sabiñanigo',
'Medina-sidonia': 'Medina sidonia',
'Alcoba de los montes': 'Alcoba',
'Sarinyena': 'Sariñena',
'Palma de mallorca': 'Palma',
'Alcali de xivert': 'Alcala de xivert',
'Santa eulalia del rio': 'Santa eularia des riu',
'Pollensa': 'Pollença',
'Vistabella del maestrazgo': 'Vistabella del maestrat',
'Lumbreras': 'Lumbreras de cameros',
'Isona i conca delli': 'Isona i conca della',
'Arti': 'Arta',
'Les coves de vinromi': 'Les coves de vinroma',
'La pobla de benifassi': 'La pobla de benifassa',
'Castellon': 'Castello de la plana',
'La laguna': 'San cristobal de la laguna',
'Figols i alinyi': 'Figols i alinya',
'Mellid': 'Melide',
'Albanyi': 'Albanya',
'Tori': 'Tora',
'Llangreu': 'Langreo',
'Chert': 'Xert',
'Sant llorens des cardassar': 'Sant llorenç des cardassar',
'Orbaitzeta': 'Orbaizeta',
'Arcas del villar': 'Arcas',
'Santa eulalia del campo': 'Santa eulalia',
'Benasal': 'Benassal',
'Zahara de la sierra': 'Zahara',
'Lobera': 'Lobeira',
'Oristi': 'Orista',
'Masanet de cabrenys': 'Maçanet de cabrenys',
'La ametlla de mar': "L'ametlla de mar",
'Guardiola de berguedi': 'Guardiola de bergueda',
'Oyarzun': 'Oiartzun',
'Las rozas': 'Las rozas de madrid',
'Arrazua-ubarrundia': 'Arratzua-ubarrundia',
'Espluga de francoli': "L'espluga de francoli",
'San lorenzo del escorial': 'San lorenzo de el escorial',
'Vilaflor': 'Vilaflor de chasna',
'Montelli i martinet': 'Montella i martinet',
'Samartin del rei aurelio': 'San martin del rey aurelio',
'Vall de gallinera': 'La vall de gallinera',
'Llusi': 'Lluça',
'Egües': 'Valle de egües/eguesibar',
'Sant carles de la rapita': 'La rapita',
'Argensola': 'Argençola',
'Masanet de la selva': 'Maçanet de la selva',
'Cassi de la selva': 'Cassa de la selva',
'Collado': 'Collado de la vera',
'Pozorrubio': 'Pozorrubio de santiago',
'Paracuellos del jarama': 'Paracuellos de jarama',
'Borredi': 'Borreda',
'Benlloch': 'Benlloc',
'Porquera': 'Porqueira',
'Bagi': 'Baga',
'Jarque': 'Jarque de moncayo',
'Legazpia': 'Legazpi',
'Sant llorens savall': 'Sant llorenç savall',
'Uharte-arakil': 'Uharte arakil',
'Salvatierra/agurain': 'Agurain/salvatierra',
'Ollo': 'Valle de ollo/ollaran',
'Brunyola': 'Brunyola i sant marti sapresa',
'La ampolla': "L'ampolla",
'Espunyola': "L'espunyola",
'Vidri': 'Vidra',
'Calonge': 'Calonge i sant antoni',
'Almazora': 'Almassora',
'Etxarri-aranatz': 'Etxarri aranatz',
'Albi': "L'albi",
'Olleria': "L'olleria",
'Sant llorens de la muga': 'Sant llorenç de la muga',
'La cisterniga': 'Cisterniga',
'Castelltersol': 'Castellterçol',
'Maldi': 'Malda',
'Pardinas': 'Pardines',
'Gavi': 'Gava',
'La aldea': "L'aldea",
'Bigues i riells': 'Bigues i riells del fai',
'Llansi': 'Llança',
'Cornelli del terri': 'Cornella del terri',
'Herbes': 'Herbers',
'Barberi de la conca': 'Barbera de la conca',
'Aleixar': "L'aleixar",
'Albages': "L'albages",
'Massanas': 'Massanes',
'Guimeri': 'Guimera',
'Ribera baja/erribera beitia': 'Erriberabeitia',
'La nou de berguedi': 'La nou de bergueda',
'Otura': 'Villa de otura',
'Villafranca de bonany': 'Vilafranca de bonany',
'Sant boi de llusanes': 'Sant boi de lluçanes',
'La alcudia': "L'alcudia",
'Calpe': 'Calp',
"La vall d'alcali": "La vall d'alcala",
'Arrankudiaga': 'Arrankudiaga-zollo',
'Capsanes': 'Capçanes',
"Llisi d'amunt": "Lliça d'amunt",
'Les masies de voltregi': 'Les masies de voltrega',
'Castrillo matajudios': 'Castrillo mota de judios',
'Vistabella de huerva': 'Vistabella',
'Espluga calva': "L'espluga calba",
'Sant jaume de frontanyi': 'Sant jaume de frontanya',
"La bisbal d'empordi": "La bisbal d'emporda",
'Albiol': "L'albiol",
"Sant llorens d'hortons": "Sant llorenç d'hortons",
'Celri': 'Celra',
'Foixi': 'Foixa',
'Sarratella': 'La serratella',
'Puigcerdi': 'Puigcerda',
'Pusol': 'Puçol',
'La lantejuela': 'Lantejuela',
'Sant vicens de castellet': 'Sant vicenç de castellet',
'Balenyi': 'Balenya',
'Serinyi': 'Serinya',
'Manjabalago': 'Manjabalago y ortigosa de rioalmar',
"La tallada d'empordi": "La tallada d'emporda",
'La escala': "L'escala",
'Abanto y ciervana': 'Abanto y ciervana-abanto zierbena',
'Castellseri': 'Castellsera',
'Esponelli': 'Esponella',
'Estella/lizarra': 'Estella-lizarra',
'Genoves': 'El genoves',
'Palau-soliti i plegamans': 'Palau-solita i plegamans',
'Adsubia': "L'atzubia",
'Savalli del comtat': 'Savalla del comtat',
'Algimia de alfara': "Algimia d'alfara",
'Placencia de las armas': 'Soraluze-placencia de las armas',
"L' ametlla del valles": "L'ametlla del valles",
"L'arbos": "L'arboç",
'Canyellas': 'Canyelles',
'Santa eulalia de ronsana': 'Santa eulalia de ronçana',
'Hospitalet de llobregat': "L'hospitalet de llobregat",
'Prats de llusanes': 'Prats de lluçanes',
'Torms': 'Els torms',
'Sant agusti de llusanes': 'Sant agusti de lluçanes',
'Valle de trapaga': 'Valle de trapaga-trapagaran',
"Bellcaire d'empordi": "Bellcaire d'emporda",
'Organyi': 'Organya',
'Urrotz': 'Urroz',
'Tortelli': 'Tortella',
'Llisi de vall': 'Lliça de vall',
'Ciurana': 'Siurana',
'Estany': "L'estany",
'Facheca': 'Fageca',
'Argentera': "L'argentera",
'Borrassi': 'Borrassa',
'Favareta': 'Favara',
'Sant vicens dels horts': 'Sant vicenç dels horts',
"El far d'empordi": "El far d'emporda",
'Polinyi de xuquer': 'Polinya de xuquer',
'Polinyi': 'Polinya',
'Santa cecilia de voltregi': 'Santa cecilia de voltrega',
'Pedret i marzi': 'Pedret i marza',
'Palleji': 'Palleja',
'Rodonyi': 'Rodonya',
'Barberi del valles': 'Barbera del valles',
'Oitz': 'Oiz',
'Sant vicens de montalt': 'Sant vicenç de montalt',
'Palomar': 'El palomar',
'Enova': "L'enova",
'Almiseri': 'Almisera',
'Ulli': 'Ulla',
'Cornelli de llobregat': 'Cornella de llobregat',
'Sant vicens de torello': 'Sant vicenç de torello',
'Masllorens': 'Masllorenç',
'Rotgli i corberi': 'Rotgla i corbera',
'Parlavi': 'Parlava',
'Real de gandia': 'El real de gandia',
'Gainza': 'Gaintza',
'Llorens del penedes': 'Llorenç del penedes',
'La armentera': "L'armentera",
'Palma de cervello': 'La palma de cervello',
'Rafelbuñol': 'Rafelbunyol',
'Sant llorens de morunys': 'Sant llorenç de morunys',
'Guadasequies': 'Guadassequies',
'Torre endomenech': "La torre d'en domenec",
'Rourell': 'El rourell',
'Benisano': 'Benissano',
'Benisuera': 'Benissuera',
'Cerdi': 'Cerda',
'Sant hipolit de voltregi': 'Sant hipolit de voltrega',
'Benirredri': 'Benirredra',
#añadimos los últimos cambios al diccionario después de revisar los municipios que no coinciden:
'El mili':'El mila',
'Alegria de alava':'Alegria-dulantzi',
'Alcocer de planes':'Alcosser',
'Jalon':'Xalo',
'Deya':'Deia',
'Ibiza':'Eivissa',
'Santa maria de corco':"L'esquirol",
'San mateo':'Sant mateu',
'Corsi':'Corça',
'Flasi':'Flaça',
'Bidegoian':'Bidania-goiatz',
'Deva':'Deba',
'Veracruz':'Beranuy',
'Marsi':'Marça',
'Roda de bari':'Roda de bera',
'Alqueria de la condesa':"L'alqueria de la comtessa",
'Exeya':'Ejea de los caballeros',
'Villanueva de castellon':'Castello',
'Pradales':'Carabias'}

#reemplazamos los nombres de municipios en la lista de superfice:
sup_2.replace(replacements, inplace = True)

#repetimos merge para ver con cuántos valores únicos nos quedamos ahora:
pop_den = pd.merge(poblacion,sup_2, how='outer', on=['municipios'])
pop_den.tail(5)

pop_den.shape

pop_den.isna().sum()

pop_den[pop_den['km2'].isna() == True]['municipios'].unique()

pop_den[pop_den['cod_ine'].isna() == True]['municipios'].unique()

#Eliminamos los municipios restantes que no aparecen en la lista del INE
pop_den.dropna(subset=['cod_ine'], inplace = True)

#completamos datos de superficie para Frontera, Ceuta y Melilla:
pop_den[(pop_den['municipios'] == 'Frontera') | (pop_den['municipios'] == 'Ceuta') | (pop_den['municipios'] == 'Melilla')]

pop_den.iloc[5754,7] = 'Caceres'
pop_den.iloc[5754,8] = 'Castilla-La Mancha'
pop_den.iloc[5754,9] = 34.57
pop_den.iloc[8161,7] = 'Ceuta'
pop_den.iloc[8161,8] = 'Ceuta'
pop_den.iloc[8161,9] = 18.5
pop_den.iloc[8162,7] = 'Melilla'
pop_den.iloc[8162,8] = 'Melilla'
pop_den.iloc[8162,9] = 12.3

pop_den.isna().sum()

pop_den.rename(columns={'provincia_x':'prov_id', 'municipio':'mun_id','cod_ine':'ine_id', 'municipios':'municipio','provincia_y':'provincia'}, inplace = True)
pop_den.head(1)

pop_den.prov_id =pop_den.prov_id.apply(lambda x: str(x).zfill(2))
pop_den.head(1)

pop_den['densidad'] = pop_den['total_2023']/pop_den['km2']
pop_den.head(5)

pop_den.mun_id = pop_den.prov_id + pop_den.mun_id

pop_den.drop(columns=['provincia','comunidad'], inplace = True)

#pop_den.to_csv('densidad_poblacion.csv', encoding='utf-8-sig', index = False)
