==> Notebooks/3_location_codes.ipynb <==
import pandas as pd
import numpy as np
import re
import seaborn as sns
import matplotlib.pyplot as plt
from bs4 import BeautifulSoup as bs
import requests

locations = pd.read_excel('../Data/Raw/loc_code.xlsx')

locations.head(2)

locations['CODAUTO'] = locations['CODAUTO'].apply(lambda x: str(x).zfill(2))

locations['CPRO'] = locations['CPRO'].apply(lambda x: str(x).zfill(2))

locations['CMUN'] = locations['CMUN'].apply(lambda x: str(x).zfill(3))

locations.rename(columns={'CODAUTO':'com_id', 'CPRO':'prov_id','CMUN':'mun_id'}, inplace = True)
locations.drop(columns=['DC'], inplace = True)

locations.mun_id = locations.prov_id + locations.mun_id
locations.head(4)

#locations.to_csv('mun_codes.csv', encoding='utf-8-sig', index = False)
