==> Notebooks/4_prov_com_codes.ipynb <==
import pandas as pd
import numpy as np
import re
import seaborn as sns
import matplotlib.pyplot as plt
from bs4 import BeautifulSoup as bs
import requests

table = pd.read_html('https://www.ine.es/daco/daco42/codmun/cod_ccaa_provincia.htm')

cods = pd.DataFrame(table[0])
cods.drop(50, axis=0, inplace = True)
cods.head(1)

cambios = pd.DataFrame({})
cambios['original'] = (cods['Comunidad Autónoma'].tolist())+(cods['Provincia'].tolist())
cambios[['c1','c2']] = cambios['original'].str.split(',', expand = True)
cambios = cambios[cambios['c2'].isna() == False]

original= cambios['original'].tolist()
clean = (cambios['c2']+' '+cambios['c1']).tolist()

clean_dict = dict(zip(original, clean))
clean_dict

cods.replace(clean_dict, inplace = True)

cods

cod_coms = cods.loc[:,['CODAUTO', 'Comunidad Autónoma']]
cod_coms.rename(columns={'CODAUTO':'com_id','Comunidad Autónoma':'comunidad'}, inplace = True)
cod_coms.drop_duplicates(subset='comunidad', keep='first', inplace = True)
#cod_coms.to_csv('cod_comundades.csv', encoding='utf-8-sig', index = False)

cod_coms.head(2)

cod_provs = cods.loc[:,['CPRO','CODAUTO','Provincia']]
cod_provs.rename(columns={'CPRO':'prov_id','CODAUTO':'com_id','Provincia':'provincia'}, inplace = True)
cod_provs.dropna(inplace = True)
cod_provs['prov_id'] = cod_provs['prov_id'].apply(lambda x: int(x))
cod_provs['com_id'] = cod_provs['com_id'].apply(lambda x: int(x))
#cod_provs.to_csv('cod_provincias.csv', encoding='utf-8-sig', index = False)

cod_provs.head(2)
