==> Notebooks/5_infection_tables.ipynb <==
import pandas as pd
import numpy as np
import re
import seaborn as sns
import matplotlib.pyplot as plt
from bs4 import BeautifulSoup as bs
import requests

densidad = pd.read_csv('../Data/clean_data/densidad_poblacion.csv')

densidad.head(1)

contagio_mun = densidad.loc[:,['densidad', 'prov_id', 'mun_id', 'municipio']]

contagio_mun.head(2)

#calculating zombie virus estimated basic reproduction number (transmision rate): infectiousnes * population density

c = 5 #constant representing infectiousness of the virus (For reference, flu is 1.3, measles is 12-18. We are picking 8 given that every human is susceptible of contracting the virus, there is no recovery and once bitten you are infected with a 100% probability.)

contagio_mun['R0'] = contagio_mun.densidad*c

contagio_mun['prob_inf'] = 1 - np.exp((-1*contagio_mun.R0*contagio_mun.densidad))

contagio_mun.head(3)

#contagio_mun.to_csv('contagio_municipios.csv', encoding='utf-8-sig', index = False)

contagio_prov = pd.read_csv('../Data/clean_data/cod_provincias.csv')

contagio_prov.head(2)

#calculamos los datos de superficie, población total y densidad por provincia a partir de los datos por municipio

contagio_prov['km2'] = contagio_prov.prov_id.apply(lambda x: sum(densidad[densidad.prov_id == x]['km2']))

contagio_prov['total_prov'] = contagio_prov.prov_id.apply(lambda x: sum(densidad[densidad.prov_id == x]['total_2023']))

contagio_prov['densidad'] = contagio_prov['total_prov'] / contagio_prov['km2']
contagio_prov.head(2)

#calculamos el índice y probablilidad de contagio por provincia

contagio_prov['R0'] = contagio_prov.densidad*c
contagio_prov['prob_inf'] = 1 - np.exp((-1*contagio_prov.R0*contagio_prov.densidad))

contagio_prov.head(3)

#repetimos el proceso por comunidad autónoma:

coms = contagio_prov.iloc[:,[0,1,3,4]]

coms.head(1)

contagio_com = pd.read_csv('../Data/clean_data/cod_comundades.csv')

contagio_com.head(1)

#población, superficie y densidad de pobalción por comunidad a partir de datos de provincia:

contagio_prov['km2'] = contagio_prov.prov_id.apply(lambda x: sum(densidad[densidad.prov_id == x]['km2']))

contagio_com['km2'] = contagio_com.com_id.apply(lambda x: sum(contagio_prov[contagio_prov.com_id == x]['km2']))

contagio_com['total_com'] = contagio_com.com_id.apply(lambda x: sum(contagio_prov[contagio_prov.com_id == x]['total_prov']))

contagio_com['densidad'] = contagio_com['total_com'] / contagio_com['km2']
contagio_com.head(2)

contagio_com.head(1)

#índice y probabilidad de contagio por comunidad:

contagio_com['R0'] = contagio_com.densidad*c
contagio_com['prob_inf'] = 1 - np.exp((-1*contagio_com.R0*contagio_com.densidad))

contagio_com.head(1)

#contagio_com.to_csv('contagio_comunidades.csv', encoding='utf-8-sig', index = False)

#quitamos la columna de com_id de la tabla de contagio por provincia:

#contagio_prov.drop(columns=['com_id'], inplace = True)

contagio_prov.head(2)

#contagio_prov.to_csv('contagio_provincias.csv', encoding='utf-8-sig', index = False)
