==> Notebooks/6_SQL.ipynb <==
import pandas as pd
import numpy as np
import re
import seaborn as sns
import matplotlib.pyplot as plt
from bs4 import BeautifulSoup as bs
import requests

mun_codes = pd.read_csv('../Data/clean_data/mun_codes.csv')
mun_codes.head(1)

prov_codes = pd.read_csv('../Data/clean_data/cod_provincias.csv')

com_codes = pd.read_csv('../Data/clean_data/cod_comundades.csv')

provs = pd.read_csv('../Data/clean_data/contagio_provincias.csv')

muns = pd.read_csv('../Data/clean_data/contagio_municipios.csv')

coms = pd.read_csv('../Data/clean_data/contagio_comunidades.csv')

#uploading data to our mysql database

from sqlalchemy import create_engine
from sqlalchemy import text
import passwords

password = passwords.password_sql

connection_string = 'mysql+pymysql://root:' + password + '@localhost:3306/zombies'

engine = create_engine(connection_string)

#códigos de comunidades autónomas:
com_codes.to_sql(name = 'cod_comunidades',
                  con = engine,
                  if_exists = 'append', 
                 index = False)

#probabilidad de contagio por comunidades autónomas:
coms.to_sql(name = 'inf_com',
                  con = engine,
                  if_exists = 'append', 
                 index = False)

#códigos de provincias:
prov_codes.to_sql(name = 'cod_provincias',
                  con = engine,
                  if_exists = 'append', 
                 index = False)

#probabilidad de contagio por provincias:
provs.to_sql(name = 'inf_prov',
                  con = engine,
                  if_exists = 'append', 
                 index = False)


#códigos de municipios:
mun_codes.to_sql(name = 'cod_municipios',
                  con = engine,
                  if_exists = 'append', 
                 index = False)


#probabilidad de contagio por municipios:
muns.to_sql(name = 'inf_muns',
                  con = engine,
                  if_exists = 'append', 
                 index = False)
